+++ CHANGELOG.md
@@ -2,14 +2,17 @@
 
 ## CHANGELOG
 
-### 3.0.0 (26-Sep-2025) 
+### 3.0.0 (03-Oct-2025)
 
 **First (attempt at a) MDM-agnostic release**
 
+### 2.5.0 (03-Oct-2025)
 - Added "System Memory" and "System Storage" capacity information (Pull Request #36; thanks again, @HowardGMac!)
 - Corrected misspelling of "Certificate" in multiple locations (Pull Request #41; thanks, @HowardGMac!)
-- Improved handling of the `checkJamfProCheckIn` and `checkJamfProInventory` functions when no relevant data is found in the jamf.log file
-- Refactored `mdmClientAvailableOSUpdates`
+- Improved handling of the `checkJamfProCheckIn` and `checkJamfProInventory` functions when no relevant data is found in the `jamf.log` file
+- Refactored `checkAvailableSoftwareUpdates` to include DDM-enforced OS Updates
+- Added error-handling for `organizationOverlayiconURL`
+- Minor Cisco VPN fixes (Pull Request #47; thanks, @HowardGMac!)
 - Introduces a new `operationMode` of "Silent" to run all checks and log results without displaying a dialog to the end-user
 
     > :warning: **Breaking Change** :warning:
@@ -17,6 +20,28 @@
     > The `operationMode` variable is now case-sensitive and the former "production" option has been renamed to "Self Service".
     > 
     > Please update any existing policies that set this variable to use: "Test", "Debug", "Self Service" or "Silent" (with initial capital letters).
+
+    <details>
+        <summary>Click to view screenshots</summary>
+        <details>
+            <summary>Script</summary>
+            <img src="images/MHC_2.5.0_Script_Options.png" alt="Settings > Computer Management > Scripts > Options > Parameter Labels > Parameter 4" width="600"/><br/>
+            Settings > Computer Management > Scripts > Options > Parameter Labels > Parameter 4<br/><br/>
+            <code>Operation Mode [ Test | Debug | Self Service | Silent ]</code>
+        </details>
+        <details>
+            <summary>Self Service Policy</summary>
+            <img src="images/MHC_2.5.0_Policy_Self_Service.png" alt="Computers > Policies > Options > Scripts > Parameter Values > Self Service" width="600"/><br/>
+            Computers > Policies > Options > Scripts > Parameter Values > <code>Self Service</code>
+        </details>
+        <details>
+            <summary>Silent Policy</summary>
+            <img src="images/MHC_2.5.0_Policy_Silent_General.png" alt="Computers > Policies > Options > General > Trigger > Custom > customTriggerName" width="600"/><br/>
+            Computers > Policies > Options > General > Trigger > Custom > <code>customTriggerName</code><br/><br/>
+            <img src="images/MHC_2.5.0_Policy_Silent_Scripts.png" alt="Computers > Policies > Options > Scripts > Parameter Values > Silent" width="600"/><br/>
+            Computers > Policies > Options > Scripts > Parameter Values > <code>Silent</code>
+        </details>
+    </details>
 
 ### 2.4.0 (20-Sep-2025)
 - Updated SSID code (thanks, ZP!)